Added an option to stop training. Added a warning window before trying to leave a training session. MainWindow.cpp is split into several files (by pages). Fixed bugs on the training page and on the model configuration page when exporting files. Small fixes in gui
Former-commit-id: b4c18cde886ffbc104a2f5c3f002ae938501d468 [formerly c446b626f241e624236d94fc67d94206f0b8e7e3]
Former-commit-id: da27fb76e9897d9e0c663cf13fadedde4bee31f5

+++ src/core/utils/BMPReader.cpp
@@ -23,7 +23,7 @@ std::vector<float> BMPDataToGrayscale(const unsigned char* bmp_data,
 	max = max - (max - min) / 20;
 	min = min + (max - min) / 20;
 	for (auto &i : grayscale_data)
-		i = std::min(255., std::max(0., (i - min) * (1. / (max - min))));
+		i = (float) std::min(255., std::max(0., (i - min) * (1. / (max - min))));
 	return grayscale_data;
 }
